Add extensions to content_repo, turn, turn_bq, survey and rapidpro.
