+++ rdw_ingestion_tools/api/rapidpro/requests/flow_starts.py
@@ -2,7 +2,8 @@ from attrs import define
 from httpx import Client
 from pandas import DataFrame
 
-from ..extensions.dataframe import concatenate
+from api import concatenate
+
 from ..extensions.httpx import get_paginated
