+++ rdw_ingestion_tools/api/survey/__init__.py
@@ -1,31 +1,9 @@
-from collections.abc import Iterator
-
 from httpx import Client
 
 from .. import config_from_env
 
 API_KEY = config_from_env("SURVEY_API_KEY")
 BASE_URL = config_from_env("SURVEY_API_BASE_URL")
-
-
-def get_paginated(
-    client: Client, url: str, start: str, max_pages: int, **kwargs: str | int
-) -> Iterator[dict]:
-    """Paginates over the custom MQR Survey API endpoint."""
-    pages = 0
-
-    params = {"updated_at_gt": start, **kwargs}
-
-    while url and pages < max_pages:
-        response = client.get(url=url, params=params)
-        response.raise_for_status()
-
-        data: dict = response.json()["results"]
-        yield from data
-
-        url = response.json()["next"]
-
-        pages += 1
 
 
 headers = {"Authorization": f"Token {API_KEY}"}
