+++ rdw_ingestion_tools/api/survey/requests/mqr.py
@@ -1,8 +1,10 @@
 from attrs import define
 from httpx import Client
-from pandas import DataFrame, concat, json_normalize
+from pandas import DataFrame
 
-from .. import get_paginated
+from api import concatenate
+
+from ..extensions.httpx import get_paginated
 
 
 @define
@@ -25,13 +27,6 @@ class MQR:
             self.client, url, start=ts, max_pages=max_pages
         )
 
-        baseline_responses = []
-        for item in baseline_generator:
-            baseline_responses.append(json_normalize(item, sep="_"))
-
-        try:
-            baseline = concat(baseline_responses)
-        except ValueError:
-            baseline = DataFrame()
+        baseline = concatenate(baseline_generator)
 
         return baseline
